+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 Future release (next version)
 =====
 
+Put unreleased changes here!
+
 Frameworks:
 - Add Services with intent filter boot_completed
 
 OS:
 - build_mpos.sh: suppress Clang 21's -Wunterminated-string-initialization so the macOS build doesn't fail compiling secp256k1
-Put unreleased changes here!
+- Create WifiBoot-, WebServerBoot- and AIOReplService
 
 0.11.2
 ======